third_party/blink/web_tests/fast/spatial-navigation/snav-single-select-expected.txt

	2	
4	
1
2
3
	6
	8	
PASS successfullyParsed is true

TEST COMPLETE

PASS gFocusedDocument.activeElement.getAttribute("id") is "8"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "2"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "6"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "4"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "start"
PASS gFocusedDocument.activeElement.getAttribute("id") is "8"
PASS gFocusedDocument.getElementById("start").selectedIndex is 0
PASS gFocusedDocument.getElementById("start").selectedIndex is 1
PASS gFocusedDocument.getElementById("start").selectedIndex is 0
